arith_ext: add integer ranges and normalised operation
