+++ lib/Dialect/ArithExt/IR/ArithExtOps.h
@@ -3,6 +3,7 @@
 
 #include "lib/Dialect/ArithExt/IR/ArithExtDialect.h"
 #include "mlir/include/mlir/IR/BuiltinOps.h"  // from @llvm-project
+#include "mlir/include/mlir/Interfaces/InferIntRangeInterface.h"  // from @llvm-project
 #include "mlir/include/mlir/Interfaces/InferTypeOpInterface.h"  // from @llvm-project
 
 #define GET_OP_CLASSES
