+++ lib/Dialect/ArithExt/IR/BUILD
@@ -11,6 +11,7 @@ cc_library(
     name = "Dialect",
     srcs = [
         "ArithExtDialect.cpp",
+        "ArithExtOps.cpp",
     ],
     hdrs = [
         "ArithExtDialect.h",
@@ -22,6 +23,7 @@ cc_library(
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:ArithDialect",
         "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:InferIntRangeInterface",
         "@llvm-project//mlir:InferTypeOpInterface",
         "@llvm-project//mlir:Support",
     ],
@@ -38,6 +40,7 @@ td_library(
     deps = [
         "@heir//lib/DRR",
         "@llvm-project//mlir:BuiltinDialectTdFiles",
+        "@llvm-project//mlir:InferIntRangeInterfaceTdFiles",
         "@llvm-project//mlir:InferTypeOpInterfaceTdFiles",
         "@llvm-project//mlir:OpBaseTdFiles",
         "@llvm-project//mlir:SideEffectInterfacesTdFiles",
